+++ store/prefix/store.go
@@ -12,9 +12,9 @@ import (
 
 var _ types.KVStore = Store{}
 
-// Store is similar with cometbft/cometbft/libs/db/prefix_db
+// Store is similar with cometbft/cometbft-db/blob/v1.0.1/prefixdb.go
 // both gives access only to the limited subset of the store
-// for convinience or safety
+// for convenience or safety
 type Store struct {
 	parent types.KVStore
 	prefix []byte
@@ -42,12 +42,12 @@ func (s Store) key(key []byte) (res []byte) {
 	return
 }
 
-// Implements Store
+// GetStoreType implements Store
 func (s Store) GetStoreType() types.StoreType {
 	return s.parent.GetStoreType()
 }
 
-// Implements CacheWrap
+// CacheWrap implements CacheWrap
 func (s Store) CacheWrap() types.CacheWrap {
 	return cachekv.NewStore(s)
 }
@@ -57,31 +57,31 @@ func (s Store) CacheWrapWithTrace(w io.Writer, tc types.TraceContext) types.Cach
 	return cachekv.NewStore(tracekv.NewStore(s, w, tc))
 }
 
-// Implements KVStore
+// Get implements KVStore
 func (s Store) Get(key []byte) []byte {
 	res := s.parent.Get(s.key(key))
 	return res
 }
 
-// Implements KVStore
+// Has implements KVStore
 func (s Store) Has(key []byte) bool {
 	return s.parent.Has(s.key(key))
 }
 
-// Implements KVStore
+// Set implements KVStore
 func (s Store) Set(key, value []byte) {
 	types.AssertValidKey(key)
 	types.AssertValidValue(value)
 	s.parent.Set(s.key(key), value)
 }
 
-// Implements KVStore
+// Delete implements KVStore
 func (s Store) Delete(key []byte) {
 	s.parent.Delete(s.key(key))
 }
 
-// Implements KVStore
-// Check https://github.com/cometbft/cometbft/blob/master/libs/db/prefix_db.go#L106
+// Iterator implements KVStore
+// Check https://github.com/cometbft/cometbft-db/blob/v1.0.1/prefixdb.go#L109
 func (s Store) Iterator(start, end []byte) types.Iterator {
 	newstart := cloneAppend(s.prefix, start)
 
@@ -98,7 +98,7 @@ func (s Store) Iterator(start, end []byte) types.Iterator {
 }
 
 // ReverseIterator implements KVStore
-// Check https://github.com/cometbft/cometbft/blob/master/libs/db/prefix_db.go#L129
+// Check https://github.com/cometbft/cometbft-db/blob/v1.0.1/prefixdb.go#L132
 func (s Store) ReverseIterator(start, end []byte) types.Iterator {
 	newstart := cloneAppend(s.prefix, start)
 
@@ -134,17 +134,17 @@ func newPrefixIterator(prefix, start, end []byte, parent types.Iterator) *prefix
 	}
 }
 
-// Implements Iterator
+// Domain implements Iterator
 func (pi *prefixIterator) Domain() ([]byte, []byte) {
 	return pi.start, pi.end
 }
 
-// Implements Iterator
+// Valid implements Iterator
 func (pi *prefixIterator) Valid() bool {
 	return pi.valid && pi.iter.Valid()
 }
 
-// Implements Iterator
+// Next implements Iterator
 func (pi *prefixIterator) Next() {
 	if !pi.valid {
 		panic("prefixIterator invalid, cannot call Next()")
@@ -156,7 +156,7 @@ func (pi *prefixIterator) Next() {
 	}
 }
 
-// Implements Iterator
+// Key implements Iterator
 func (pi *prefixIterator) Key() (key []byte) {
 	if !pi.valid {
 		panic("prefixIterator invalid, cannot call Key()")
@@ -168,7 +168,7 @@ func (pi *prefixIterator) Key() (key []byte) {
 	return
 }
 
-// Implements Iterator
+// Value implements Iterator
 func (pi *prefixIterator) Value() []byte {
 	if !pi.valid {
 		panic("prefixIterator invalid, cannot call Value()")
@@ -177,7 +177,7 @@ func (pi *prefixIterator) Value() []byte {
 	return pi.iter.Value()
 }
 
-// Implements Iterator
+// Close implements Iterator
 func (pi *prefixIterator) Close() error {
 	return pi.iter.Close()
 }
@@ -192,7 +192,7 @@ func (pi *prefixIterator) Error() error {
 	return nil
 }
 
-// copied from github.com/cometbft/cometbft/libs/db/prefix_db.go
+// copied from github.com/cometbft/cometbft-db/blob/v1.0.1/prefixdb.go
 func stripPrefix(key, prefix []byte) []byte {
 	if len(key) < len(prefix) || !bytes.Equal(key[:len(prefix)], prefix) {
 		panic("should not happen")