+++ store/v2/mock/mockgen.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+mockgen_cmd="mockgen"
+$mockgen_cmd -package mock -destination ./db_mock.go -source ./types.go